+++ lib/core/AppManager.ts
@@ -298,7 +298,7 @@ export default class AppManager {
     * @param callback - The effect callback
     * @param dependencies - Array of state keys to listen for
     */
-   public static useEffect(callback: () => Function|undefined, dependencies: string[]|null = null) {
+   public static useEffect(callback: () => void | (() => void), dependencies: string[]|null = null) {
       RuntimeManager.registerEffect(callback, dependencies);
    }