+++ lib/core/RuntimeManager.ts
@@ -50,7 +50,7 @@ export default class RuntimeManager {
     * @param {Function} callback - The effect callback
     * @param {Array<string>} dependencies - Array of state keys to listen for
     */
-   public static registerEffect(callback: () => Function|undefined, dependencies: string[]|null = null): void {
+   public static registerEffect(callback: () => void | (() => void), dependencies: string[]|null = null): void {
       // --- Run immediately (mount) ---
       const cleanup = callback();