+++ src/script/types/types/RuntimeInterfaces.d.ts
@@ -1,7 +1,7 @@
 export type EffectType = {
-    callback: Function;
+    callback: () => void | (() => void);
     dependencies: Array<string> | null;
-    cleanup: Function | null;
+    cleanup: (() => void) | null;
 };
 export interface EventObject {
     beforeload: ((payload: EventPayload) => void)[];